abstracts parser in preparation for handling multiple sources
